Implement Multi-User Functionality

+++ backend/controllers/workoutController.js
@@ -3,7 +3,9 @@ const mongoose = require('mongoose')
 
 // get all workouts
 const getWorkouts = async (req, res) => {
-  const workouts = await Workout.find({}).sort({createdAt: -1})
+  const user_id = req.user._id
+
+  const workouts = await Workout.find({ user_id }).sort({createdAt: -1})
 
   res.status(200).json(workouts)
 }
@@ -11,12 +13,14 @@ const getWorkouts = async (req, res) => {
 // get single workout
 const getWorkout = async (req, res) => {
   const { id } = req.params
+  const user_id = req.user._id
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return res.status(404).json({error: 'No such workout'})
   }
 
-  const workout = await Workout.findById(id)
+  // const workout = await Workout.findById(id)
+  const workout = await Workout.findOne({ _id: id, user_id: user_id }); // makes sure that user id also matches
 
   if (!workout) {
     return res.status(404).json({error: 'No such workout'})
@@ -44,7 +48,8 @@ const createWorkout = async (req, res) => {
   }
 
   try {
-    const workout = await Workout.create({title, load, reps})
+    const user_id = req.user._id
+    const workout = await Workout.create({title, load, reps, user_id})
     res.status(200).json(workout)
   } catch (error) {
     res.status(400).json({error: error.message})
@@ -54,12 +59,13 @@ const createWorkout = async (req, res) => {
 // delete a workout
 const deleteWorkout = async (req, res) => {
   const { id } = req.params
+  const user_id = req.user._id
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return res.status(404).json({error: 'No such workout'})
   }
 
-  const workout = await Workout.findOneAndDelete({_id: id})
+  const workout = await Workout.findOneAndDelete({_id: id, user_id: user_id})
 
   if (!workout) {
     return res.status(404).json({error: 'No such workout'})
@@ -71,12 +77,13 @@ const deleteWorkout = async (req, res) => {
 // update a workout
 const updateWorkout = async (req, res) => {
   const { id } = req.params
+  const user_id = req.user._id
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return res.status(404).json({error: 'No such workout'})
   }
 
-  const workout = await Workout.findOneAndUpdate({_id: id}, {
+  const workout = await Workout.findOneAndUpdate({_id: id, user_id: user_id}, {
     ...req.body
   })
